refactor: 모달 매니저의 범용성을 높임.
- openModal 함수는 모달 컨텐트 컴포넌트를 인자로 받아, 렌더링 및 모달 관련 처리만 담당하도록 수정. (어떤 모달인지는 돈케어)
- openFormModal.tsx 함수를 별도로 만들고, 특정 프로퍼티를 만족하는 컴포넌트를 인자로 넘겨줌. (FormModalContent)
