+++ src/components/index.ts
@@ -1,3 +1,2 @@
 export * from "./AccessableModal";
 export * from "./FormModalContent";
-export * from "./ModalManager";
