+++ src/modal/ModalManager.tsx
@@ -1,11 +1,11 @@
-import type { FormInput } from "../types";
+import type { BaseModalProps, FormInput } from "../types";
 import { createRoot } from "react-dom/client";
-import { AccessibleModal } from "./AccessableModal";
-import { FormModalContent } from "./FormModalContent";
+import { AccessibleModal } from "../components";
 
-// TODO: 모달 컴포넌트를 인자로 받아서 렌더링하도록 수정
-export const openFormModal = (
-  element: HTMLButtonElement,
+export const openModal = (
+  props: Pick<BaseModalProps, "titleId" | "descriptionId">,
+  ModalContent: React.FC<BaseModalProps>,
+  focusReturnElement: HTMLButtonElement,
 ): Promise<FormInput | null> => {
   // 모달용 포탈 생성 + DOM에 붙이기
   const portalRoot =
@@ -25,20 +25,20 @@ export const openFormModal = (
       document.body.classList.remove("overflow-hidden");
 
       // 클릭한 요소에 포커스 돌려주기
-      element.focus();
+      focusReturnElement.focus();
     };
 
     // 폼 모달을 createRoot 를 이용해 렌더링.
     // createPortal 은 JSX 영역에서만 동작하는 듯.
     createRoot(portalRoot).render(
       <AccessibleModal
-        titleId="modal-title-id"
-        descriptionId="modal-desc-id"
+        titleId={props.titleId}
+        descriptionId={props.descriptionId}
         onClose={handleClose}
       >
-        <FormModalContent
-          titleId="modal-title-id"
-          descriptionId="modal-desc-id"
+        <ModalContent
+          titleId={props.titleId}
+          descriptionId={props.descriptionId}
           onClose={handleClose}
         />
       </AccessibleModal>,
