+++ src/modal/index.ts
@@ -1,0 +1,2 @@
+export * from "./ModalManager";
+export * from "./openFormModal";
