+++ src/modal/openFormModal.tsx
@@ -1,0 +1,9 @@
+import { openModal } from "./ModalManager";
+import { FormModalContent } from "../components";
+
+export const openFormModal = (trigger: HTMLButtonElement) =>
+  openModal(
+    { titleId: "form-modal-title", descriptionId: "form-modal-desc" },
+    FormModalContent,
+    trigger,
+  );
