+++ src/types.ts
@@ -2,3 +2,9 @@ export interface FormInput {
   email: string;
   // ... 기타 필드
 }
+
+export interface BaseModalProps {
+  titleId: string;
+  descriptionId?: string;
+  onClose: (data: any) => void;
+}
